# src_backend/controller/GroupsController.ts
import {
  Controller,
  Route,
  Tags,
  Get,
  Post,
  Delete,
  Path,
  Body,
  Security,
} from "tsoa";
import { groupService } from "../service/groupService";
import { GroupResponse, CreateGroupRequest, AddMemberRequest } from "../model/types";

@Route("groups")
@Tags("Groups")
export class GroupsController extends Controller {
  @Get()
  @Security("jwt")
  public async list(): Promise<GroupResponse[]> {
    const groups = await groupService.list();
    return groups as GroupResponse[];
  }

  @Post()
  @Security("jwt")
  public async create(@Body() body: CreateGroupRequest): Promise<GroupResponse> {
    const group = await groupService.create(body.name, body.description);
    return group as GroupResponse;
  }

  @Get("{id}")
  @Security("jwt")
  public async getById(@Path() id: number): Promise<GroupResponse> {
    const group = await groupService.getById(id);
    if (!group) {
      this.setStatus(404);
      throw new Error("Group not found");
    }
    return group as GroupResponse;
  }

  @Post("{id}/members")
  @Security("jwt")
  public async addMember(
    @Path() id: number,
    @Body() body: AddMemberRequest,
  ): Promise<{ message: string }> {
    await groupService.addUser(body.userId, id);
    return { message: "Member added" };
  }

  @Delete("{id}/members/{userId}")
  @Security("jwt")
  public async removeMember(
    @Path() id: number,
    @Path() userId: number,
  ): Promise<{ message: string }> {
    await groupService.removeUser(userId, id);
    return { message: "Member removed" };
  }
}